// src/detectStaleStacks/lib/infra-account-library/models/ous/OuName.ts
export enum OuName {
  Root = 'Root',
  Workloads = 'Workloads',
  GOVUKOnce = 'GOV.UK Once',
  GOVUKOnceInfra = 'GOV.UK Once Infra',
  GOVUKOnceKeyManagement = 'GOV.UK Once Key Management',
  GOVUKOnceSandbox = 'GOV.UK Once Sandbox',
  GOVUKOnceWorkloads = 'GOV.UK Once Workloads',
  GOVUKOnceDev = 'GOV.UK Once Dev',
  GOVUKOnceStag = 'GOV.UK Once Staging',
  GOVUKOnceProd = 'GOV.UK Once Production',
}

export default OuName;